refactor(arangodb): move editState struct and pipeline steps to feature_annotation_pipeline.go
The editState struct and associated pipeline steps are moved to
feature_annotation_pipeline.go to improve code organization and
maintainability. This change centralizes the logic related to
updating feature annotations, making it easier to understand and
modify the update process.
