+++ internal/repository/arangodb/feature_annotation.go
@@ -247,227 +247,6 @@ func (fann *featureAnnoRepo) processPublicationType(
 	return nil
 }
 
-// editState holds the state during the annotation editing pipeline.
-type editState struct {
-	fann        *featureAnnoRepo
-	doc         *feature.FeatureAnnotationUpdate
-	txr         *manager.TransactionHandler
-	origDoc     *model.FeatureAnnotationDoc
-	updatedDoc  *model.FeatureAnnotationDoc
-	updateQuery string
-	Err         error
-}
-
-// stepFetchOriginalDoc fetches the original document to be updated.
-func stepFetchOriginalDoc(state *editState) *editState {
-	if state.Err != nil {
-		return state
-	}
-
-	var err error
-	state.origDoc, err = state.fann.GetFeatureAnnotation(state.doc.Id)
-	if err != nil {
-		if repository.IsAnnotationNotFound(err) {
-			state.Err = err // Preserve specific error type
-		} else {
-			state.Err = fmt.Errorf(
-				"error fetching original document: %w",
-				err,
-			)
-		}
-	}
-
-	return state
-}
-
-// stepBeginTransaction starts a transaction for the annotation update.
-func stepBeginTransaction(state *editState) *editState {
-	if state.Err != nil {
-		return state
-	}
-
-	state.txr, state.Err = state.fann.database.BeginTransaction(
-		context.Background(),
-		&manager.TransactionOptions{
-			WriteCollections: []string{
-				state.fann.feature.Name(),
-				state.fann.pub.Name(),
-				state.fann.edge.Name(),
-			},
-		})
-
-	if state.Err != nil {
-		state.Err = fmt.Errorf("error beginning transaction: %w", state.Err)
-	}
-
-	return state
-}
-
-// stepUpdateDocFields updates the document fields with new values.
-func stepUpdateDocFields(state *editState) *editState {
-	if state.Err != nil {
-		return state
-	}
-
-	// Update fields
-	updateBasicFields(state.origDoc, state.doc)
-	if state.doc.Attributes != nil {
-		updateAttributes(state.origDoc, state.doc.Attributes)
-	}
-
-	// Prepare the update query
-	state.updateQuery = fmt.Sprintf(
-		"UPDATE @doc WITH @data IN %s RETURN NEW",
-		state.fann.feature.Name(),
-	)
-
-	return state
-}
-
-// stepExecuteUpdate executes the update query and reads the result.
-func stepExecuteUpdate(state *editState) *editState {
-	if state.Err != nil {
-		return state
-	}
-
-	result, err := state.txr.DoRun(
-		state.updateQuery,
-		map[string]interface{}{
-			"doc":  state.origDoc.Key,
-			"data": state.origDoc,
-		},
-	)
-	if err != nil {
-		state.Err = fmt.Errorf("error updating feature annotation: %w", err)
-
-		return state
-	}
-
-	// Read the updated result
-	state.updatedDoc = &model.FeatureAnnotationDoc{}
-	if err := result.Read(state.updatedDoc); err != nil {
-		state.Err = fmt.Errorf("error reading updated document: %w", err)
-	}
-
-	return state
-}
-
-// stepHandlePublications processes publications if needed.
-func stepHandlePublications(state *editState) *editState {
-	if state.Err != nil {
-		return state
-	}
-
-	// Skip if no attributes
-	if state.doc.Attributes == nil {
-		return state
-	}
-
-	// Process Pubmed IDs
-	if !collection.IsEmpty(state.doc.Attributes.Pubmed) {
-		if err := state.fann.processPublicationType(
-			state.txr,
-			state.updatedDoc,
-			state.doc.Attributes.Pubmed,
-			"pubmed",
-		); err != nil {
-			state.Err = fmt.Errorf(
-				"error processing pubmed publications: %w",
-				err,
-			)
-
-			return state
-		}
-	}
-
-	// Process DOI publications
-	if !collection.IsEmpty(state.doc.Attributes.Publications) {
-		if err := state.fann.processPublicationType(
-			state.txr,
-			state.updatedDoc,
-			state.doc.Attributes.Publications,
-			"doi",
-		); err != nil {
-			state.Err = fmt.Errorf("error processing DOI publications: %w", err)
-
-			return state
-		}
-	}
-
-	return state
-}
-
-// stepCommitTransaction commits the transaction.
-func stepCommitTransaction(state *editState) *editState {
-	// If there's no error, commit the transaction
-	if state.Err == nil {
-		if err := state.txr.Commit(); err != nil {
-			state.Err = fmt.Errorf("error committing transaction: %w", err)
-		}
-
-		return state
-	}
-	// If there's an error but no transaction, just return the state
-	if state.txr == nil {
-		return state
-	}
-
-	// If there's an error and a transaction, abort it
-	abortErr := state.txr.Abort()
-	if abortErr != nil {
-		state.Err = fmt.Errorf(
-			"%v, also failed to abort transaction: %w",
-			state.Err,
-			abortErr,
-		)
-	}
-
-	return state
-}
-
-// stepValidateInput validates the input document before proceeding.
-func stepValidateInput(state *editState) *editState {
-	if state.Err != nil {
-		return state
-	}
-
-	// Check for required fields
-	if state.doc.Id == "" {
-		state.Err = fmt.Errorf("feature ID is required")
-
-		return state
-	}
-
-	if state.doc.UpdatedBy == "" {
-		state.Err = fmt.Errorf("updater email is required")
-
-		return state
-	}
-
-	return state
-}
-
-// stepRefreshDocumentState retrieves the final document state after all
-// modifications.
-func stepRefreshDocumentState(state *editState) *editState {
-	if state.Err != nil {
-		return state
-	}
-
-	// Get the latest document state with all changes (including publication edges)
-	updatedDoc, err := state.fann.GetFeatureAnnotation(state.doc.Id)
-	if err != nil {
-		state.Err = fmt.Errorf("error retrieving final document state: %w", err)
-
-		return state
-	}
-
-	// Update the state with the refreshed document
-	state.updatedDoc = updatedDoc
-
-	return state
-}
-
 // EditFeatureAnnotation updates an existing feature annotation with new values.
 // It handles updating basic properties, attributes, and related publications.
 //
